+++ SpartaTextRPG2/SpartaTextRPG2/Item.cs
@@ -11,13 +11,13 @@ namespace SpartaTextRPG2
         public string Name { get; set; }
         public string Type { get; set; }
 
-        public int Power { get; set; }
+        public float Power { get; set; }
         public int Price { get; set; }
         public string Description;
         public bool PlayerEquipped { get; set; }
         public string Owner { get; set; }
 
-        public Item(string name, string type, int power, string description, bool playerEquipped, string owner)
+        public Item(string name, string type, float power, string description, bool playerEquipped, string owner)
         {
             Name = name;
             Type = type;
